Add support for OAuth Authorization Code PKCE flow
Added new auth type 'login' which can be used to integrate with
UiPath's external applications. Users need to set up a
Non-Confidential application with a redirect url pointing to the
CLI http server. When the CLI is used for the first time,
a browser window will be opened where the user can login with
the normal credentials (username/password, microsoft login, google
login, LinkedIn or custom SSO. After the login, the browser is
redirected to the local HTTP web server which has been started
by the CLI.

An authorisation code flow with PKCE is performed and
the CLI retrieves a bearer token which is cached and used for calls
to the UiPath services.

- Added new oauth_authenticator which starts the HTTP server in the
  background and opens the browser to enter the login credentials

- Added html page which is shown after the login was performed
  successfully

- Extended the uipathcli config command to support configuring
  multiple auth types: credentials or login