Add --wait flag to support waiting for conditions
Added a new --wait flag which allows users to specify a JMESPath
expression on the response body. The uipathcli keeps calling the
executor and retries the operation until the provided condition
evaluates to true.

This feature allows customers of the CLI to write a sync call which
waits for some backend operation to be carried out instead of polling
manually in a loop.

e.g.

The following command adds the DocumentUnderstanding service to a tenant:
uipath oms tenant update-tenant --organization-guid "..." \
                                --tenant-guid "..." \
                                --services "du=true"

But the operation to add a service to the tenant is asynchronous and can
take some time, but instead of polling, users can now write:

uipath oms tenant get-tenant --organization-guid "..." \
                             --tenant-guid "..." \
                             --wait "tenantServiceInstances[?serviceType == 'du'].status == 'Enabled'"

The CLI outputs:
Condition is not met yet. Waiting...

Until the operation timed out or the field status for the du serviceType
is "Enabled".

The default timeout is 30s but can be adjusted by providing the
--wait-timeout flag, e.g.

uipath oms tenant get-tenant --organization-guid "..." \
                             --tenant-guid "..." \
			     --wait "tenantServiceInstances[?serviceType == 'du'].status == 'Enabled'" \
                             --wait-timeout 300