Split tests by plugin command and move commands into separate folders
There are quite a lot of studio plugin tests. Refactored the tests and
moved them into a separate test file per command. This makes it easier
to have an overview of the existing tests for a particular command.

Also split up the large studio plugin package into separate folders for
each command.